Feat: Set articles and comments visibility
* Those resources can now be public, private or archived.
* The articles index show the current number of public articles.

[Use concerns to DRY]
Instead of using duplicated logic on article and comment models, we rely on a
concept abstraction about resources visibility to isolate this logic in a
concern (module) included on both models.

[Buggy articles count]
Since the previously created articles have no defined visibility status, they
are not counted by the public_count method.